bldprogs/VBoxEditElf: Started implementing the code to generate an ELF shared object linker stub from the internal representation [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110438 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/bldprogs/VBoxEditElf.cpp
@@ -236,7 +236,7 @@ static uint32_t elfEditImgGetStrIdxInStrTab(PELFEDITSTUBIMG pStubImg, const char
     PELFEDITSTR pStr = (PELFEDITSTR)RTStrSpaceGet(&pStubImg->StrSpace, pszString);
     AssertRelease(pStr && pStr->offStrTab > 0);
 
-    return pStr->offStrTab;
+    return (uint32_t)pStr->offStrTab;
 }
